refactor(web): CustomRoomServer
